+++ engineFoxtrot/Engine/Video/VideoFacade.cpp
@@ -2,8 +2,8 @@
 #include "VideoFacade.h"
 
 #include <SDL.h>
-#include "../../SDL2/include/SDL_image.h"
-#include "../../SDL2/include/SDL_ttf.h"
+#include <SDL_image.h>
+#include <SDL_ttf.h>
 
 #undef main
 
@@ -110,7 +110,7 @@ void VideoFacade::renderCopy(Object& object)
 	//generate image 
 	Uint32 ticks = SDL_GetTicks();
 	Uint32 seconds = ticks / sprite.getAnimationDelay();
-	int leftpos = sprite.getLeftPos(seconds);
+	float leftpos = sprite.getLeftPos(seconds);
 	int top = 0;
 
 	//generate rectangele for selecting 1 image of a full sprite
@@ -118,7 +118,7 @@ void VideoFacade::renderCopy(Object& object)
 	//top = amount of pixels of the top (sprites are renderd of the top to bottom
 	//width = amount of pixels of the with of 1 image
 	//height = amount of pixels of the height of 1 image
-	SDL_Rect rect{ leftpos, top, sprite.getWidth(), sprite.getHeight() };
+	SDL_Rect rect{ (int)leftpos, top, (int)sprite.getWidth(), (int)sprite.getHeight() };
 
 	//update collision box 
 	if (!object.getScalable()) {
@@ -128,10 +128,10 @@ void VideoFacade::renderCopy(Object& object)
 
 	//generate stratch of image
 	SDL_Rect destination;
-	destination.x = object.getPositionX();
-	destination.y = object.getPositionY() - object.getHeight();
-	destination.w = object.getWidth();
-	destination.h = object.getHeight();
+	destination.x = (int)object.getPositionX();
+	destination.y = (int)object.getPositionY() - (int)object.getHeight();
+	destination.w = (int)object.getWidth();
+	destination.h = (int)object.getHeight();
 	SDL_RenderCopyEx(renderer, textureMap[sprite.getTextureID()], &rect, &destination, object.getRotation(), NULL, SDL_FLIP_NONE);
 }
